pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minpaku-video"
version = "0.1.0"
description = "民泊経営チャンネル向けスライド動画制作パイプライン"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "elevenlabs>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "mutagen>=1.47.0",
    "httpx>=0.27.0",
]

[project.scripts]
minpaku-video = "minpaku_video.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/minpaku_video"]
